requirements.txt
langchain>=0.3.0,<0.4.0
langchain-anthropic>=0.3.0,<0.4.0
web3>=7.0.0,<8.0.0
aiohttp>=3.9.0,<4.0.0
pandas>=2.2.0,<3.0.0
numpy>=1.26.0,<2.0.0
redis>=5.0.0,<6.0.0
python-dotenv>=1.0.0,<2.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
